+++ rc-application/rc-app-mall-biz/src/main/java/com/rc/cloud/app/distributor/infrastructure/persistence/po/DistributorDetailDO.java
@@ -23,9 +23,9 @@ public class DistributorDetailDO {
      * id
      */
     @TableId(type = IdType.ASSIGN_ID)
-    private Integer id;
+    private Long id;
 
-    private Integer distributorId;
+    private Long distributorId;
 
     private String distributorDetail;
 }
